third_party/WebKit/LayoutTests/fast/loader/show-only-one-beforeunload-dialog-expected.txt

CONFIRM NAVIGATION: This is beforeunload from the top level frame.
CONSOLE MESSAGE: Blocked attempt to show multiple 'beforeunload' confirmation panels for a single navigation.
CONSOLE MESSAGE: Blocked attempt to show multiple 'beforeunload' confirmation panels for a single navigation.
